Rename seed.js to cryptoPairs.js for clarity
