+++ seed/cryptoPairs.js
